refactor: Extract config and flag parser
